+++ packages/api/src/resolvers/queries/events/eventsCounts.ts
@@ -1,8 +1,13 @@
 import { log } from '@unchainedshop/logger';
 import { Root, Context } from '@unchainedshop/types/api.js';
-import { EventQuery } from '@unchainedshop/types/events.js';
 
-export default async function eventsCount(root: Root, params: EventQuery, { modules, userId }: Context) {
+type CountEventsParams = Parameters<Context['modules']['events']['count']>['0'];
+
+export default async function eventsCount(
+  root: Root,
+  params: CountEventsParams,
+  { modules, userId }: Context,
+) {
   log(`query eventsCount  queryString: ${params.queryString}  types: ${params.types}  ${userId}`, {
     userId,
   });