+++ packages/api/src/resolvers/type/work-types.ts
@@ -1,11 +1,15 @@
 import { Context } from '@unchainedshop/types/api.js';
 import { Work as WorkType } from '@unchainedshop/types/worker.js';
+import { buildObfuscatedFieldsFilter } from '@unchainedshop/utils';
 
 type HelperType<P, T> = (work: WorkType, params: P, context: Context) => T;
 
 export interface WorkHelperTypes {
   status: HelperType<never, string>;
   type: HelperType<never, string>;
+  result: HelperType<never, any>;
+  input: HelperType<never, any>;
+  error: HelperType<never, any>;
   original: HelperType<never, Promise<WorkType>>;
 }
 
@@ -21,5 +25,17 @@ export const Work: WorkHelperTypes = {
   original: async (obj, _, { modules }) => {
     if (!obj.originalWorkId) return null;
     return modules.worker.findWork({ workId: obj.originalWorkId });
+  },
+
+  input: (obj, _, { options }: Context) => {
+    return buildObfuscatedFieldsFilter(options.worker?.blacklistedVariables)(obj.input);
+  },
+
+  result: (obj, _, { options }: Context) => {
+    return buildObfuscatedFieldsFilter(options.worker?.blacklistedVariables)(obj.result);
+  },
+
+  error: (obj, _, { options }: Context) => {
+    return buildObfuscatedFieldsFilter(options.worker?.blacklistedVariables)(obj.error);
   },
 };