+++ packages/core-events/src/db/EventsCollection.ts
@@ -1,8 +1,15 @@
 import type { mongodb } from '@unchainedshop/mongodb';
-import { Event } from '@unchainedshop/types/events.js';
 import { buildDbIndexes } from '@unchainedshop/mongodb';
+import { TimestampFields } from '@unchainedshop/types/common.js';
 
 const TWO_DAYS_SEC = 172800;
+
+export type Event = {
+  _id?: string;
+  type: string;
+  context?: Record<string, unknown>;
+  payload?: Record<string, unknown>;
+} & TimestampFields;
 
 export const EventsCollection = async (db: mongodb.Db) => {
   const Events = db.collection<Event>('events');