+++ packages/core-events/src/events-index.ts
@@ -1,1 +1,3 @@
+export type { Event } from './db/EventsCollection.js';
+
 export { configureEventsModule } from './module/configureEventsModule.js';