+++ packages/types/index.d.ts
@@ -106,6 +106,7 @@ import {
   IWorkerDirector,
   WorkerModule,
   WorkerSchedule,
+  WorkerSettingsOptions,
   WorkStatus as WorkerStatusType,
 } from './worker.js';
 import { ModuleInput } from './core.js';
@@ -290,7 +291,7 @@ declare module '@unchainedshop/core-users' {
 
 declare module '@unchainedshop/core-warehousing' {
   function configureWarehousingModule(
-    params: ModuleInput<Record<string, never>>,
+    params: ModuleInput<WorkerSettingsOptions>,
   ): Promise<WarehousingModule>;
 
   const WarehousingDirector: IWarehousingDirector;
@@ -300,7 +301,7 @@ declare module '@unchainedshop/core-warehousing' {
 }
 
 declare module '@unchainedshop/core-worker' {
-  function configureWorkerModule(params: ModuleInput<Record<string, never>>): Promise<WorkerModule>;
+  function configureWorkerModule(params: ModuleInput<WorkerSettingsOptions>): Promise<WorkerModule>;
 
   const WorkerDirector: IWorkerDirector;
   const WorkStatus: typeof WorkerStatusType;